Diplom/requirements.txt
asgiref==3.2.10
atomicwrites==1.4.0
attrs==20.3.0
colorama==0.4.4
coverage==5.5
Django==3.1.2
django-filter==2.4.0
djangorestframework==3.12.1
iniconfig==1.1.1
model-bakery==1.2.1
packaging==20.9
Pillow==8.1.0
pluggy==0.13.1
psycopg2-binary==2.8.6
py==1.10.0
pyparsing==2.4.7
pytest==6.2.3
pytest-cov==2.11.1
pytest-django==4.2.0
pytz==2021.1
sqlparse==0.4.1
toml==0.10.2
